Programmatic container management and dependency scanning (#474)
Part one of a planned series of PRs addressing #473. This PR:

- Implements programmatic container management using the Wave API: each
container is built according to a YAML specification in `containers` by
the `bin/build_wave_container.py` script, which also updates the
corresponding path in `configs/containers.config`.
(`bin/build_wave_containers.py` is a convenience script that runs
`bin/build_wave_container.py` for all YAML files in the directory).
- Replaces former Docker Hub containers with Wave containers built in
this way; all module tests pass.
- Implements dependency scanning with Trivy via
`bin/scan_containers.py`, and integrates this into Github Actions. (Note
that, currently, the Github Actions test is expected to fail.)
- Deletes now-redundant Dockerfiles for Docker Hub containers.

This PR does _not_ actually update the versions of any dependencies, and
so is compatible with a point release. In a future PR, I'll update
dependencies and try to get the Trivy test to pass.