Implement similarity based duplicate marking
The current version of the pipeline only has alignment-based duplicate marking,
where two reads are considered duplicates if they align to the same portion of
their identified reference genomes.  This is very computationally efficient,
but it misses duplicates in cases where the reads are assigned to different
genomes.  Apply the duplication detection logic Evan wrote for Outward Assembly
to identify additional duplicates.

This is not integrated into the pipeline at all: in its current state the user
needs to manually run this as a post-processing step.  To get this into
production quickly we'll start with this as part of
https://github.com/securebio/nao-mgs-partner-reports and then later if we like
the output we'll most likely make this part of DOWNSTREAM.  See
https://twist.com/a/197793/ch/619193/t/7439978/ for discussion.

First step of fixing https://github.com/securebio/nao-mgs-workflow/issues/341